Basic shade model is working
This reads an STL model from stdin, generates a POV-Ray model that
tests sun visibility to a particular test point for every minute over
the course of a year, and plots the result.

It has some basic performance optimizations, like a simple caching
layer around the POV-Ray modeling to make it fast to iterate on the
visualization part.

The usability could definitely use some work. A lot of things are
hard-coded that should be options. I'd like to add a "partial shade"
model. I'd like to change the plot to show sun intensity, not just
"light or dark", and also generate a plot of totals over the day.
Running POV-Ray on even a simple test model takes about a minute; if
that gets worse I might have to do my own 3-D math.